Update CONTRIBUTING.md
---
title: Contributing to an open source for good project
intro: How to find an open source for good project
versions:
  fpt: '*'
---
## Contribute to open source for good projects

For Good First Issue curates social impact and civic tech open source projects that are looking for contributors and lists issues on those projects that are tagged with the `help wanted` or `good first issue` labels.

Open source maintainers are always looking to get more people involved, but it can be challenging to become a contributor. For Good First Issue lowers the barrier for future contributions to social and civic good projects - and this is why it exists.

### How to find and contribute to open source

1. Navigate to [For Good First Issue](https://forgoodfirstissue.github.com/).
1. Search for a project of your interest based on the Sustainable Development Goals.
1. Select a project and navigate to the project's repository.
1. Review the project's Contribution Guidelines and README to get started

Please note that these projects are curated by the community and will be regularly reviewed.